Simplify Stylix component return value
I tried at one point to prevent an extra component wrapper from appearing in the React dev tools, but it turns out that old-style class components can't do this; it's not possible to accurately detect a class vs. a constructor function. So you'll just have to live with the extra element in the dev tools for now.